refactor: update default formatter and clean up client ID handling
