+++ package.json
@@ -47,7 +47,8 @@
     "swagger-ui-dist": "^5.20.6",
     "swagger-ui-react": "^5.20.6",
     "tempo-devtools": "^2.0.99",
-    "vaul": "^1.1.2"
+    "vaul": "^1.1.2",
+    "zod": "^3.24.2"
   },
   "devDependencies": {
     "@types/node": "^20",
