+++ src/app/(auth)/sign-in/page.tsx
@@ -2,12 +2,11 @@
 
 import { signInAction } from "@/app/actions";
 import FormMessage from "@/components/form-message";
-import Navbar from "@/components/navbar";
 import { SubmitButton } from "@/components/submit-button";
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
-import Link from "next/link";
 import { Message } from "@/types/message";
+import Link from "next/link";
 
 
 interface LoginProps {
@@ -22,7 +21,7 @@ export default function SignInPage({ searchParams }: LoginProps) {
 
   return (
       <>
-        <Navbar />
+        {/* <Navbar /> */}
         <div className="flex min-h-screen flex-col items-center justify-center bg-background px-4 py-8">
           <div className="w-full max-w-md rounded-lg border border-border bg-card p-6 shadow-sm">
             <form className="flex flex-col space-y-6" action={signInAction}>
