+++ src/app/(auth)/sign-up/page.tsx
@@ -1,12 +1,11 @@
-import { FormMessage, Message } from "@/components/form-message";
+import { signUpAction } from "@/app/actions";
+import FormMessage, { Message } from "@/components/form-message";
 import { SubmitButton } from "@/components/submit-button";
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
+import { UrlProvider } from "@/components/url-provider";
 import Link from "next/link";
 import { SmtpMessage } from "../smtp-message";
-import { signUpAction } from "@/app/actions";
-import Navbar from "@/components/navbar";
-import { UrlProvider } from "@/components/url-provider";
 
 export default async function Signup(props: {
   searchParams: Promise<Message>;
@@ -22,13 +21,15 @@ export default async function Signup(props: {
 
   return (
     <>
-      <Navbar />
+      {/* <Navbar /> */}
       <div className="flex min-h-screen flex-col items-center justify-center bg-background px-4 py-8">
         <div className="w-full max-w-md rounded-lg border border-border bg-card p-6 shadow-sm">
           <UrlProvider>
             <form className="flex flex-col space-y-6">
               <div className="space-y-2 text-center">
-                <h1 className="text-3xl font-semibold tracking-tight">Sign up</h1>
+                <h1 className="text-3xl font-semibold tracking-tight">
+                  Sign up
+                </h1>
                 <p className="text-sm text-muted-foreground">
                   Already have an account?{" "}
                   <Link
