--- src/app/client/[clientId]/page.tsx
@@ -1,10 +1,0 @@
-import { redirect } from "next/navigation";
-
-export default function ClientIdRedirect({
-  params,
-}: {
-  params: { clientId: string };
-}) {
-  // Redirect to the clients route with standardized parameter
-  return redirect(`/clients/${params.clientId}`);
-}
