+++ src/app/client/[id]/page.tsx
@@ -3,8 +3,8 @@ import { redirect } from "next/navigation";
 export default function ClientIdRedirect({
   params,
 }: {
-  params: { clientId: string };
+  params: { id: string };
 }) {
   // Redirect to the clients route with standardized parameter
-  return redirect(`/clients/${params.clientId}`);
+  return redirect(`/clients/${params.id}`);
 }
